requirements.txt
cachetools==5.3.1
click==8.1.7
cloudpickle==2.2.1
cuda-python==12.2.0
cudf-cu12==23.8.0
cuml-cu12==23.8.0
cupy-cuda12x==12.2.0
Cython==3.0.0
dask==2023.7.1
dask-cuda==23.8.0
dask-cudf-cu12==23.8.0
distributed==2023.7.1
duckdb==0.8.1
fastrlock==0.8.1
fsspec==2023.6.0
importlib-metadata==6.8.0
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
llvmlite==0.40.1
locket==1.0.0
MarkupSafe==2.1.3
msgpack==1.0.5
numba==0.57.1
numpy==1.24.4
nvidia-ml-py3==7.352.0
nvtx==0.2.6
packaging==23.1
pandas==1.5.3
partd==1.4.0
protobuf==4.24.1
psutil==5.9.5
pyarrow==11.0.0
pylibraft-cu12==23.8.0
pynvml==11.4.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
raft-dask-cu12==23.8.0
rmm-cu12==23.8.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sklearn==0.0.post7
sortedcontainers==2.4.0
tblib==2.0.0
threadpoolctl==3.2.0
toolz==0.12.0
tornado==6.3.3
treelite==3.2.0
treelite-runtime==3.2.0
typing_extensions==4.7.1
ucx-py-cu12==0.33.0
urllib3==2.0.4
zict==3.0.0
zipp==3.16.2
